+++ tests/TestCase.php
@@ -157,7 +157,7 @@ class TestCase extends \Orchestra\Testbench\TestCase
         ]);
 
         // Subscribe test user to plan
-        $this->testUser->newSubscription('main', $this->testPlanBasic);
+        $this->testUser->newSubscriptionPlan('main', $this->testPlanBasic);
 
         $this->testUser->subscription('main')->features()->create(['tag' => 'social_cat_profiles', 'name' => 'Social profiles available for your cat', 'value' => 10, 'sort_order' => 10]);
     }